# Speedup Rational.Str