By generating the fractional part two characters at a time instead of
one.

For a FatRat that ends up being a 1571 char string, time before was 2.75s, time after is 1.9s.
For a FatRat that ends up being a 3139 char string, time before was 16.2s, time after is 12.6s.

Rakudo builds ok and passes `make m-test m-spectest`.