@AlexDaniel 2018-02-23
Heh, but why not 3 or 4 characters at a time? :joy: What's so magical about 2?