@zoffixznet 2018-02-23
👍 Thanks!

> What's so magical about 2?

You need a check for roundness and the lower numbers would still have to be handled because `$precision` might not chunk up precisely. Notice the original 1-char loop is still in the code and then 2-char loop has a roundness check. Adding more chars would mean more loops and more checks, which gets progressively messier.